requirements.txt
asgiref==3.3.1
Django==3.0.5
django-debug-toolbar==3.2
mysqlclient==2.0.3
python-decouple==3.4
pytz==2020.5
sqlparse==0.4.1
